Fixed IOC name validation check
The IOC device generator script would take a user input name as a
command line argument. In the event in incorrect name (i.e., longer
than 8 chars) was input the script would promt for a new valid
name. However, the script would still use the original (incorrect)
name when creating git repositories.

This change moves the validation of the name to the constructor and
only updates the IOC name once it's passed the checks. The test for
the invalid long name was also updated.